+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### 12.26.1
 
+* `[events]` Improved unknown events handling
+* `[events]` Added usage examples
 * Error check moved at the beginning of every test
 
 ### 12.26.0